fix(restructure): restructure and separate different parts into different directory
