+++ src/Background/Reducers/VideoManager.js
