+++ src/Content/DetectVideoFinished.js
@@ -1,4 +1,4 @@
-import Config from "./Config";
+import Config from "../Config";
 export default class DetectVideoFinished {
     finishCallback;
     interval;
