+++ src/Popup/Popup.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import {render} from 'react-dom';
+import render from 'react-dom';
 import {Store} from 'react-chrome-redux';
 import App from './App';
-import Config from './Config';
+import Config from '../Config';
 import './Popup.css';
 
 const proxyStore = new Store({
